pyproject.toml
[project]
name = "realitycheck-ai"
version = "0.2.0"
description = "AI-powered code quality analyzer"
authors = [{name = "Ayush"}]
dependencies = ["rich", "click","google-genai"]

[project.scripts]
realitycheck = "realitycheck.cli.main:main"